// src/api/orval/model/settingResponseApiResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { SettingResponse } from "./settingResponse";

export interface SettingResponseApiResponse {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  data?: SettingResponse;
}