// src/api/orval/model/getVersionDatabaseResponseApiResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { GetVersionDatabaseResponse } from "./getVersionDatabaseResponse";

export interface GetVersionDatabaseResponseApiResponse {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  data?: GetVersionDatabaseResponse;
}